requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
alembic>=1.12.0
pydantic>=2.5.0
pydantic[email]>=2.5.0
python-jose[cryptography]>=3.3.0
asyncpg>=0.29.0
orjson>=3.9.0
greenlet>=3.0.0
starlette>=0.27.0